Conflicts:
	include/drivers/OPH_IMPORTESDM2_operator.h
	include/drivers/OPH_IMPORTNC2_operator.h
	include/oph_nc_library.h
	src/drivers/Makefile.am
	src/drivers/OPH_CONCATESDM_operator.c
	src/drivers/OPH_IMPORTESDM2_operator.c
	src/drivers/OPH_IMPORTESDM_operator.c
	src/drivers/OPH_IMPORTNC_operator.c
	src/oph_nc_library.c

+++ src/drivers/OPH_FS_operator.c
@@ -1002,7 +1002,7 @@ int task_execute(oph_operator_struct * handle)
 							}
 
 							char filenames[jj][OPH_COMMON_BUFFER_LEN], *start = tbuffer, *save_pointer = NULL;
-							int jjc = jj;
+							size_t jjc = jj;
 							for (ii = jj = 0; ii < jjc; ++ii, start = NULL) {
 								filename = strtok_r(start, OPH_SEPARATOR_PARAM, &save_pointer);
 								lstat(filename, &file_stat);
